remove unused lines of code
