--- src/app/constants/constants.ts
@@ -1,11 +1,0 @@
-enum InstanceStatus{
-  READY="ready",
-  ACCEPTED="accepted",
-  PREPARING="preparing",
-  PROVISIONING="provisioning",
-  FAILED="failed"
-}
-
-export {
-  InstanceStatus
-}
